requirements.txt
requests==2.31.0
aiohttp==3.9.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
gspread==6.0.0
google-api-python-client==2.115.0
google-auth-oauthlib==1.2.0
groq==0.4.2
playwright==1.41.2
pdfkit==1.0.0
markdown==3.5.2
streamlit
pandas
